+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-31T10:51:55.522Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.521Z")
     r"""The date and time the resource was last updated"""