+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1320,7 +1320,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-31T10:52:06.511Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.844Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1372,7 +1372,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-31T10:52:06.510Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.843Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1424,7 +1424,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-31T10:52:06.507Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.843Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1498,5 +1498,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-31T10:51:55.200Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.201Z")
     r"""The date and time the resource was last updated"""