+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8WYASBTZ6P19ENH87K0KP7E"
+        "01K94HF0G8WP87VBX2KWVFB6P8"
     )
     r"""The unique identifier of the data source"""