+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -215,5 +215,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-31T10:51:55.522Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.521Z")
     r"""The date and time the resource was last updated"""